fixes #167

+++ vripper-gui/src/main/kotlin/me/vripper/gui/model/settings/ConnectionSettingsModel.kt
@@ -1,7 +1,6 @@
 package me.vripper.gui.model.settings
 
 import javafx.beans.property.SimpleIntegerProperty
-import javafx.beans.property.SimpleLongProperty
 import tornadofx.getValue
 import tornadofx.setValue
 
@@ -12,8 +11,8 @@ class ConnectionSettingsModel {
     val maxTotalThreadsProperty = SimpleIntegerProperty()
     var maxTotalThreads: Int by maxTotalThreadsProperty
 
-    val timeoutProperty = SimpleLongProperty()
-    var timeout: Long by timeoutProperty
+    val timeoutProperty = SimpleIntegerProperty()
+    var timeout: Int by timeoutProperty
 
     val maxAttemptsProperty = SimpleIntegerProperty()
     var maxAttempts: Int by maxAttemptsProperty
